requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
albumentations==1.3.0
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
ensemble-boxes==1.0.9
filelock==3.9.0
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
lit==15.0.7
llvmlite==0.40.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numba==0.57.0
numpy==1.24.1
opencv-contrib-python==4.4.0.46
opencv-python==4.5.5.64
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.3.0
protobuf==4.23.2
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
sentry-sdk==1.24.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.112
urllib3==1.26.13
wandb==0.15.3
zipp==3.15.0
